Refactor middlewares user endpoint
