+++ middlewares/users/controllers/deleteAll.js
@@ -1,0 +1,11 @@
+const User = require('../model')
+
+const deleteAll = async (req, res, next) => {
+  const report = await User.deleteMany({})
+  res.status(200).send({
+    message: 'All users deleted',
+    data: { deletedUsersCount: report.deletedCount },
+  })
+}
+
+module.exports = deleteAll
