+++ middlewares/users/controllers/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  getUsers: require('./getUsers'),
+  getById: require('./getById'),
+  getUserProfile: require('./getUserProfile'),
+  updateUserProfile: require('./updateUserProfile'),
+  adminUpdateUser: require('./adminUpdateUser'),
+  deleteById: require('./deleteById'),
+  deleteAll: require('./deleteAll'),
+}
